+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.9.42] - 2023-02-02 14:41:00
+### Changed
+- better handling of product images
+### Added
+- Getimage endpoint
+
 ## [4.9.41] - 2023-01-02 14:41:00
 ### Changed
 - better handling of commas in select attributes